feat: Make S3 file resources leaner & refactor
