+++ dynamicio/handlers/s3/__init__.py
@@ -1,2 +1,3 @@
 """S3 handlers for dynamicio."""
-from dynamicio.handlers.s3.file import S3CsvResource, S3HdfResource, S3JsonResource, S3ParquetResource
+from dynamicio.handlers.s3.file import S3CsvResource, S3JsonResource, S3ParquetResource
+from dynamicio.handlers.s3.hdf import S3HdfResource
