+++ dynamicio/handlers/s3/hdf.py
@@ -5,8 +5,42 @@ import uuid
 from contextlib import contextmanager
 from typing import IO, Generator, Optional
 
+import boto3  # type: ignore
 import pandas as pd  # type: ignore
 import tables  # type: ignore
+from pydantic import Field
+
+from dynamicio import utils
+from dynamicio.handlers.s3.contexts import s3_named_file_reader, s3_reader, s3_writer
+from dynamicio.handlers.s3.file import BaseS3Resource, hdf_lock
+
+
+class S3HdfResource(BaseS3Resource):
+    """S3 Resource for HDF files.
+
+    Attributes:
+        pickle_protocol: The pickle protocol to use when writing to HDF. Default is 4, which covers python 3.4+.
+    """
+
+    pickle_protocol: int = Field(4, ge=0, le=5)
+
+    _file_read_method = staticmethod(pd.read_hdf)  # type: ignore
+    _file_write_method = None  # type: ignore
+
+    def _resource_read(self) -> pd.DataFrame:
+        if self.no_disk_space:
+            with s3_reader(boto3.client("s3"), s3_bucket=self.bucket, s3_key=str(self.path)) as fobj:  # type: ignore
+                if (result := HdfIO().load(fobj)) is not None:  # type: ignore
+                    return result
+        with s3_named_file_reader(boto3.client("s3"), s3_bucket=self.bucket, s3_key=str(self.path)) as target_file:
+            with hdf_lock:
+                return self._file_read_method(target_file.name, **self.kwargs)  # type: ignore
+
+    def _resource_write(self, df: pd.DataFrame) -> None:
+        with s3_writer(boto3.client("s3"), s3_bucket=self.bucket, s3_key=str(self.path)) as fobj, utils.pickle_protocol(
+            protocol=self.pickle_protocol
+        ):
+            HdfIO().save(df, fobj, **self.kwargs)  # pylint: disable=E1101
 
 
 class InMemStore(pd.io.pytables.HDFStore):
